refactor: extract search state
